fix: update annotation integration test to use provisioned dashboard and move newTestContext to shared utility
